+++ bin/schieber
@@ -2,10 +2,10 @@
 
 import sys, signal, argparse, logging
 
-from schieber.player.cli_player import CliPlayer
-from schieber.player.challenge_player.challenge_player import ChallengePlayer
-from schieber.player.greedy_player.greedy_player import GreedyPlayer
-from schieber.player.random_player import RandomPlayer
+from schieber.players.cli_player import CliPlayer
+from schieber.players.challenge_player.challenge_player import ChallengePlayer
+from schieber.players.greedy_player.greedy_player import GreedyPlayer
+from schieber.players.random_player import RandomPlayer
 from schieber.tournament import Tournament
 
 
@@ -46,11 +46,11 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='CLI schieber', )
     parser.add_argument('-p', '--points', dest='points', type=int, help='Tournament points')
     parser.add_argument('-f', '--first_opponent', dest='first_opponent', type=int,
-                        help='Chose your first opponent. (1: Random player, 2: Greedy player, 3: Challenge player)')
+                        help='Chose your first opponent. (1: Random players, 2: Greedy players, 3: Challenge players)')
     parser.add_argument('-s', '--second_opponent', dest='second_opponent', type=int,
-                        help='Chose your second opponent. (1: Random player, 2: Greedy player, 3: Challenge player)')
+                        help='Chose your second opponent. (1: Random players, 2: Greedy players, 3: Challenge players)')
     parser.add_argument('-t', '--teammate', dest='teammate', type=int,
-                        help='Chose your opponents. (1: Random player, 2: Greedy player, 3: Challenge player)')
+                        help='Chose your opponents. (1: Random players, 2: Greedy players, 3: Challenge players)')
     parser.set_defaults(points=1500, first_opponent=1, second_opponent=1, teammate=1)
     args = parser.parse_args()
