+++ schieber/players/base_player.py
@@ -47,7 +47,7 @@ class BasePlayer:
 
     def allowed_cards_with_hand_cards(self, state: GameState, hand_cards):
         """
-        Returns the cards on the hand of the player which he/she is allowed to play in the current state according to the rules
+        Returns the cards on the hand of the players which he/she is allowed to play in the current state according to the rules
         :param hand_cards:
         :param state:
         :return:
