+++ schieber/players/challenge_player/challenge_player.py
@@ -1,8 +1,8 @@
 import random
 
 from schieber.game import GameState
-from schieber.player.base_player import BasePlayer
-from schieber.player.challenge_player.strategy.jass_strategy import JassStrategy
+from schieber.players.base_player import BasePlayer
+from schieber.players.challenge_player.strategy.jass_strategy import JassStrategy
 from schieber.card import Card
 from schieber.rules.trumpf_rules import trumpf_allowed
