+++ schieber/players/challenge_player/strategy/card_counter.py
@@ -1,13 +1,13 @@
 from math import floor
 
 from schieber.rules.deck import Deck
-from schieber.player.challenge_player.strategy.flags.doesnt_habe_card_flag import DoesntHaveCardFlag
-from schieber.player.challenge_player.strategy.flags.falied_to_serve_suit_flag import FailedToServeSuitFlag
-from schieber.player.challenge_player.strategy.flags.previously_had_stich_flag import PreviouslyHadStichFlag
-from schieber.player.challenge_player.strategy.flags.suit_verworfen_flag import SuitVerworfenFlag
-from schieber.player.challenge_player.strategy.mode.bottom_up_mode import *
-from schieber.player.challenge_player.strategy.mode.top_down_mode import *
-from schieber.player.challenge_player.strategy.mode.trumpf_color_mode import *
+from schieber.players.challenge_player.strategy.flags.doesnt_habe_card_flag import DoesntHaveCardFlag
+from schieber.players.challenge_player.strategy.flags.falied_to_serve_suit_flag import FailedToServeSuitFlag
+from schieber.players.challenge_player.strategy.flags.previously_had_stich_flag import PreviouslyHadStichFlag
+from schieber.players.challenge_player.strategy.flags.suit_verworfen_flag import SuitVerworfenFlag
+from schieber.players.challenge_player.strategy.mode.bottom_up_mode import *
+from schieber.players.challenge_player.strategy.mode.top_down_mode import *
+from schieber.players.challenge_player.strategy.mode.trumpf_color_mode import *
 from schieber.rules.stich_rules import stich_rules
 from schieber.stich import PlayedCard
 from schieber.rules.suit import Suit
