+++ schieber/players/challenge_player/strategy/flags/doesnt_habe_card_flag.py
