+++ schieber/players/challenge_player/strategy/flags/previously_had_stich_flag.py
