+++ schieber/players/challenge_player/strategy/flags/suit_verworfen_flag.py
