+++ schieber/players/challenge_player/strategy/mode/mode.py
