+++ schieber/players/challenge_player/strategy/mode/trumpf_color_mode.py
@@ -1,4 +1,4 @@
-from schieber.player.challenge_player.strategy.mode.mode import Mode
+from schieber.players.challenge_player.strategy.mode.mode import Mode
 from schieber.helpers.game_helper import *
 from schieber.rules.trumpf import Trumpf
 from schieber.card import from_string_to_card
