+++ schieber/players/greedy_player/trumpf_decision.py
