+++ schieber/stich.py
@@ -7,7 +7,7 @@ Stich = namedtuple('Stich', ['player', 'played_cards', 'trumpf'])
 def played_card_dict(played_card):
     """
     Returns a dictionary containing:
-    - the player who played the card
+    - the players who played the card
     - the played card
     :param played_card:
     :return:
@@ -21,7 +21,7 @@ def played_card_dict(played_card):
 def stich_dict(stich):
     """
     Returns a dictionary of the stich containing:
-    - the id of the player who is winner of the stich
+    - the id of the players who is winner of the stich
     - the trumpf
     - the cards played in the stich
     :param stich:
