+++ tests/player/challenge_player/strategy/mode/test_bottom_up_mode.py
@@ -2,7 +2,7 @@ import pytest
 
 from schieber.card import Card
 from schieber.rules.suit import Suit
-from schieber.player.challenge_player.strategy.mode.bottom_up_mode import BottomUpMode
+from schieber.players.challenge_player.strategy.mode.bottom_up_mode import BottomUpMode
 
 
 @pytest.fixture
