+++ tests/player/test_base_player.py
@@ -1,4 +1,4 @@
-from schieber.player.base_player import BasePlayer
+from schieber.players.base_player import BasePlayer
 from itertools import count
