+++ tests/test_dealer.py
@@ -1,6 +1,6 @@
 import pytest
 
-from schieber.player.random_player import RandomPlayer
+from schieber.players.random_player import RandomPlayer
 
 from schieber.dealer import Dealer
