+++ tests/test_team.py
@@ -1,6 +1,6 @@
 import pytest
 
-from schieber.player.random_player import RandomPlayer
+from schieber.players.random_player import RandomPlayer
 from schieber.team import Team
